added edit button on detail page
