+++ ui/components/__init__.py
@@ -1,5 +1,6 @@
 from oarepo_ui.resources.config import TemplatePageUIResourceConfig
 from oarepo_ui.resources.resource import TemplatePageUIResource
+from flask import redirect
 
 
 class ComponentsResourceConfig(TemplatePageUIResourceConfig):
@@ -10,4 +11,14 @@ class ComponentsResourceConfig(TemplatePageUIResourceConfig):
 
 def create_blueprint(app):
     """Register blueprint for this resource."""
-    return TemplatePageUIResource(ComponentsResourceConfig()).as_blueprint()
+    blueprint = TemplatePageUIResource(ComponentsResourceConfig()).as_blueprint()
+
+    # TODO: Hacky solution to avoid error in invenio-communities where link is hardcoded
+    # they promised to correct this soon
+    app.add_url_rule(
+        "/me/communities/",
+        endpoint="invenio_app_rdm_users.communities",
+        view_func=lambda: redirect("/me/communities/"),
+    )
+
+    return blueprint
