+++ ui/documents/semantic-ui/templates/records_dashboard/search/ResultsGridItem.jsx
@@ -1,2 +1,2 @@
-import NRDocsResultsGridItem from "../../documents/search/ResultsGridItem";
+import NRDocsResultsGridItem from "@js/documents/search/ResultsGridItem";
 export default NRDocsResultsGridItem;
