+++ ui/documents/templates/semantic-ui/documents/Sidebar.jinja
@@ -1,5 +1,11 @@
 {#def metadata, record #}
-
+{% if record.links.edit_html%}
+  <section class="ui segment">
+    <SidebarLink icon="pencil" href={record.links.edit_html}>
+      {{_('Edit')}}
+    </SidebarLink>
+  </section>
+{% endif %}
 <section aria-label="{{_('metadata/objectIdentifiers.label')}}" class="ui segment">
   <h2 class="ui small header detail-sidebar-header">{{_('metadata/objectIdentifiers.label')}}</h2>
   <div class="ui link list">
